add modelio UML archive
modelio projects are made of hundreds of small files. For each minimal
change you have always to commit several file. It does not make much sense
to have this. Also because form the diff you do not get anything.

So, just commit the zip file.
Import the ZIP with modelio, make your change, export the zip and commit

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>